+++ modules/counter_12/llm_counter_12.v
@@ -16,5 +16,4 @@ module counter_12(
             end
         end
     end
-
 endmodule